refactor: splited `indicators-api.h`
